Fix Wrapper hang on Discord rate limit.
Discord.Net appears to have a bug where it fails to send a request when hitting rate limits and wait to timeout. Turns out when changing status the Wrapper was waiting for the request to be sent to Discord before allowing more messages to be sent. I have stopped the Wrapper from waiting on Discord now. I have also stopped waiting on Discord request to finish in the server manager.